WIP: Add REST endpoint for bulk update events
